Editing the tlslib to have better docstrings.

+++ lib/tlslib.py
@@ -5,98 +5,90 @@ from charmhelpers.core import hookenv
 from charmhelpers.core import unitdata
 
 
-def server_cert(destination_directory, key_path=None):
+def server_cert(directory, key_path=None):
     """
-    Copy server cert to destination_directory
+    Store the server certificate and server key in the destination directory.
 
-    @param string destination_directory dest dir for server cert
-    @param string key_path optional source key path
+    :param string directory: The directory to save the server certificate.
+    :param string key_path: The optional path to the source server key.
     """
 
     # Must remove the path characters from the local unit name.
     path_name = hookenv.local_unit().replace('/', '_')
 
-    # Optional key_path for unit tests
+    # When not specified create the server key path.
     if not key_path:
         server_key_path = 'easy-rsa/easyrsa3/pki/private/{0}.key'.format(
-                         path_name)
+            path_name)
     else:
         server_key_path = key_path
 
-    #Save the server certificate from unitdata to dest_dir
-    _save_certificate(destination_directory, 'server')
-    # Copy the unitname.key to dest_dir/server.key
-    _copy_key(destination_directory, 'server', server_key_path)
+    # Save the server certificate from unitdata to directory.
+    _save_certificate(directory, 'server')
+    # Copy the unitname.key to directory/server.key
+    _copy_key(directory, 'server', server_key_path)
 
 
-def client_cert(destination_directory, cert_path=None, key_path=None):
+def client_cert(directory, cert_path=None, key_path=None):
     """
-    Copy client cert to destination_directory
+    Copy the client certificate and client key to the destination directory.
 
-    @param string destination_directory dest dir for client cert
-    @param string key_path optional source key path
-    @param string cert_path optional source cert path
+    :param string directory: The directory to save the client certificate.
+    :param string cert_path: The optional path to the client certificate.
+    :param string key_path: The optional path to the source client key.
     """
 
-    # Optional cert path for unit tests 
+    # When not specified create the client certificate path.
     if not cert_path:
         client_cert_path = 'easy-rsa/easyrsa3/pki/issued/client.crt'
     else:
         client_cert_path = cert_path
 
-    # Optional key path for unit tests 
+    # When not specified create the client certificate path.
     if not key_path:
         client_key_path = 'easy-rsa/easyrsa3/pki/private/client.key'
     else:
         client_key_path = key_path
 
-    # Check for directory existence
-    if not os.path.isdir(destination_directory):
-        os.makedirs(destination_directory)
-        os.chmod(destination_directory, 0o770)
+    # Ensure the destination directory exists.
+    if not os.path.isdir(directory):
+        os.makedirs(directory)
+        os.chmod(directory, 0o770)
 
-    # Destination path for client.cert.
-    webapp_client_cert_path = os.path.join(destination_directory, 'client.crt')
+    # Create the destination path for client certificate.
+    webapp_client_cert_path = os.path.join(directory, 'client.crt')
 
-    # Check for client cert
+    # Check for client certificate
     if os.path.isfile(client_cert_path):
         # Copy the client.crt to dest_dir/client.crt
         copy2(client_cert_path, webapp_client_cert_path)
         # Store the path to the key in unitdata
         unitdata.kv().set('client-cert-path', webapp_client_cert_path)
-
-    # Destination path for client.key.
-    webapp_client_key_path = os.path.join(destination_directory, 'client.key')
-
-    # If it exists, copy it over
-    # TODO: Make this more robust, i.e. what happens
-    # if the key does not exist?
-    if os.path.isfile(client_key_path):
-        # Copy the client.key to dest_dir/client.key
-        copy2(client_key_path, webapp_client_key_path)
-        # Store the path to the key in unitdata
-        unitdata.kv().set('client-key-path', webapp_client_key_path)
+    # Call the method to copy the client key.
+    _copy_key(directory, 'client', client_key_path)
+    # Create the destination path for client key.
+    webapp_client_key_path = os.path.join(directory, 'client.key')
+    # Set the destination path for the client key on the unitdata.
+    unitdata.kv().set('client-key-path', webapp_client_key_path)
 
 
 def ca(directory, cert_path=None):
     """
-    When the Certificate Authority is available, copy the CA from the
-    /usr/local/share/ca-certificates/<service_name>.crt to the proper directory.
+    Copy the CA from the source to the destination directory. The tls layer
+    installs the CA on all the peers in /usr/loca/share/ca-certificates/.
 
-    @param string directory dest dir for crt
-    @param string cert_path optional source cert path
+    :param string directory: The directory to store the ca.crt file.
+    :param string cert_path: The optional path to the source CA certificate.
     """
 
-    # Normally the CA is just on the leader, but the tls layer installs the
-    # CA on all systems in the /usr/local/share/ca-certificates directory.
-    # Optional cert_path for unit test
+    # When not specified create the path to the default location.
     if not cert_path:
         ca_path = '/usr/local/share/ca-certificates/{0}.crt'.format(
                   hookenv.service_name())
     else:
         ca_path = cert_path
 
-    # Ensure the dest_dir exists.
+    # Ensure the destination directory exists.
     if not os.path.isdir(directory):
         os.makedirs(directory)
         os.chmod(directory, 0o770)
@@ -105,40 +97,44 @@ def ca(directory, cert_path=None):
     destination_ca_path = os.path.join(directory, 'ca.crt')
     if os.path.isfile(ca_path):
         copy2(ca_path, destination_ca_path)
+    else:
+        print('The CA file {0} does not exist.'.format(ca_path))
 
 
 def _copy_key(directory, prefix, key_path):
     """
     Copy the key from the easy-rsa/easyrsa3/pki/private directory to the
     specified directory.
-    
-    @param string directory dest dir for key
-    @param string prefix prefix for key
-    @param string key_path source optional key path
-    """
 
-    # Ensure the dest_dir exists.
+    :param string directory:The destination directory to store the key.
+    :param string prefix: The prefix to name the key file prefix.key.
+    :param string key_path: The path to the source key to copy.
+    """
+    # Ensure the destination directory exists.
     if not os.path.isdir(directory):
         os.makedirs(directory)
         os.chmod(directory, 0o770)
-
     # The key is not in unitdata it is in the local easy-rsa directory.
     key_name = '{0}.key'.format(prefix)
     # The key should be copied to this directory.
     destination_key_path = os.path.join(directory, key_name)
-    # Copy the key file from the local directory to the destination.
-    copy2(key_path, destination_key_path)
+    if os.path.isfile(key_path):
+        # Copy the key file from the local directory to the destination.
+        copy2(key_path, destination_key_path)
+    else:
+        print('The key file {0} does not exist.'.format(key_path))
 
 
 def _save_certificate(directory, prefix):
     """
-    Get the certificate from the charm unitdata, and write it to the proper
-    directory. The parameters are: destination directory, and prefix to use
-    for the key and certificate name.
+    Get the certificate from the charm unitdata, and write it to the
+    desination directory.
 
-    @param string directory dest dir to save cert
-    @param string prefix prefix for cert
+    :param string directory: The destination directory to save the certificate.
+    :param string prefix: The prefix used to look up the certificate in the
+    unitdata and to name the destination file prefix.crt.
     """
+    # Ensure the destination directory exists.
     if not os.path.isdir(directory):
         os.makedirs(directory)
         os.chmod(directory, 0o770)
